Refactor name to robotdataprocess
