+++ .github/workflows/python_test.yml
@@ -32,7 +32,7 @@ jobs:
       uses: snickerbockers/submodules-init@v4
     - name: Test with unittest & evaluate test coverage
       run: |
-        coverage run --source rosbag_manip -m unittest discover tests/ -v
+        coverage run --source robotdataprocess -m unittest discover tests/ -v
     - name: Coveralls GitHub Action (Upload Coverage Report)
       uses: coverallsapp/github-action@v2.3.6
       with:
