+++ src/robotdataprocess/__init__.py
@@ -1,5 +1,4 @@
-from .command_line import CmdLineInterface
-from .conversion_utils import *
+from .CmdLineInterface import CmdLineInterface
 from .data_types.ImageData import ImageData
 from .data_types.ImuData import ImuData
 from .data_types.OdometryData import OdometryData
