--- src/rosbag_manip/__main__.py
@@ -1,19 +1,0 @@
-from .command_line import CmdLineInterface
-import typer
-
-# Setup command line access
-app = typer.Typer()
-
-# Set yaml as a subcommand
-@app.command()
-def yaml(yaml_path: str):
-    """
-    Create a rosbag manipulator from a YAML file, and then run the manipulation process.
-
-    Args:
-        yaml_path (str) : Path to the YAML file containing the manipulation configuration.
-    """
-    manipulator = CmdLineInterface.from_yaml(yaml_path)
-
-if __name__ == "__main__":
-    app()
