+++ tests/test_ImuData.py
@@ -1,9 +1,9 @@
 import numpy as np
 import os
 from pathlib import Path
-from rosbag_manip import CoordinateFrame
-from rosbag_manip.data_types.ImuData import ImuData
-from rosbag_manip.rosbag.Ros2BagWrapper import Ros2BagWrapper
+from robotdataprocess import CoordinateFrame
+from robotdataprocess.data_types.ImuData import ImuData
+from robotdataprocess.rosbag.Ros2BagWrapper import Ros2BagWrapper
 import unittest
 
 class TestImuData(unittest.TestCase):
